Funcionalidades completas

+++ Users/Users/Controllers/UsersController.cs
@@ -27,7 +27,7 @@ namespace Users.Controllers
         public ActionResult<User> PostUser([FromBody] User user)
         {
             var created = _service.PostUser(user);
-            return CreatedAtAction(nameof(GetUsers), new { }, created);
+            return CreatedAtAction(nameof(GetById), new { id = created.Id }, created);
         }
 
         [HttpPatch("{id:guid}")]
@@ -42,8 +42,26 @@ namespace Users.Controllers
         public IActionResult DeleteUser(Guid id)
         {
             // DELETE /users/{id}
+            // Verificar si existe para poder dar un error 404 si no
+            var user = _service.GetUserById(id);
+            if (user == null)
+            {
+                return NotFound(new { error = true, message = "El usuario no existe." });
+            }
+
             _service.DeleteUser(id);
-            return NoContent();
+            return Ok(new
+            {
+                success = true,
+                message = $"Usuario con ID {id} eliminado correctamente."
+            });
+        }
+
+        [HttpGet("{id:guid}")]
+        public ActionResult<User> GetById(Guid id)
+        {
+            var user = _service.GetUserById(id);
+            return user == null ? NotFound() : Ok(user);
         }
 
     }
